+++ src/main/groovy/io/seqera/wave/controller/ContainerController.groovy
@@ -519,7 +519,7 @@ class ContainerController {
                 targetImage,
                 digest,
                 null, // <-- use null to copy all container tags
-                buildConfig.workspaceBucket,
+                buildConfig.buildWorkspace,
                 configJson,
                 scanId,
                 Instant.now(),
