+++ src/main/resources/application.yml
@@ -51,10 +51,12 @@ wave:
   server:
       url: "${WAVE_SERVER_URL:`http://localhost:9090`}"
   build:
-    buildkit-image: "moby/buildkit:v0.21.1-rootless"
+    workspace: "${WAVE_WORKSPACE_BUCKET}"
+    buildkit-image: "cr.seqera.io/public/wave/buildkit:77ae226fc11d42eb"
+    buildkit-image-arm64: "cr.seqera.io/public/wave/buildkit:9812468499720d11"
     singularity-image: "public.cr.seqera.io/wave/singularity:v4.2.1-r4"
-    repo: "195996028523.dkr.ecr.eu-west-1.amazonaws.com/wave/build/dev"
-    cache: "195996028523.dkr.ecr.eu-west-1.amazonaws.com/wave/build/cache"
+    repo: "hrma017/dev"
+    cache: "hrma017/cache"
     public-repo: "community.wave.seqera.io"
     # note changing this should be matching the micronaut.server.idle-timeout (see above)
     timeout: 900s
